Refactor appendFromBlend by making proper usage of the context manager.
While testing I noticed that the string list `names` in the previous version changed it's type at runtime to `bpy.types.ShaderNodeTree` and I couldn't explain why, there was no reassignment. Turns out that that context manager `with ... as ...` expression replaces the references in the list from pointing to strings, but to the actual datablocks that it imported. While it caused me quite some trouble to understand this behavior, it ultimately leads to much, much cleaner code here, as Blender already did already did all of the name - asset tracking I tried to do manually before.